% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{cov_from_graph}
\alias{cov_from_graph}
\title{Covariance matrix of observed variables}
\usage{
cov_from_graph(g, m, paths)
}
\arguments{
\item{g}{An igraph object that is a tree. It is assumed that the first m nodes correspond to oberseved nodes. 
Type 1 indicates that a node is observed. Should be set via \code{V(g)$type==1}.
It is assumed that \code{V(g)$var} is the variance of the observed variables and that \code{E(g)$corr} represents the edge correlations.}

\item{m}{Integer, number of observed nodes.}

\item{paths}{Nested list with the paths between all nodes. 
Should be computed with the function \code{\link{get_paths}}.}
}
\value{
Covarianz matrix.
}
\description{
Given a tree and parameters (variance and edge correlations) the covariance matrix of the corresponding multivariate Gaussian distribution is computed.
}
\examples{
vertices <- data.frame(name=seq(1,8), type=c(rep(1,5), rep(2,3))) # 1=observed, 2=latent
edges <- data.frame(from=c(1,2,3,4,5,6,7), to=c(8,8,6,6,7,7,8))
tree <- graph_from_data_frame(edges, directed=FALSE, vertices=vertices)
plot(tree)

# Set parameters
E(tree)$corr = rep(0.7,7)
V(tree)$var = rep(1,8)

# Compute all paths
paths <- get_paths(tree)

# Call function
cov_from_graph(tree, 5, paths)
}
