% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{random_combs}
\alias{random_combs}
\title{Randomly sample tuples from a set of integers}
\usage{
random_combs(n, k, nr)
}
\arguments{
\item{n}{Integer, determines the length of the set 1:n to select the k-tuples from.}

\item{k}{Integer, length of the tuples.}

\item{nr}{Integer, number of k-tuples being sampled.}
}
\value{
Matrix with \code{nr} rows and \code{k} columns. Each row is a k-tuple of elements from 1:n.
}
\description{
Randomly sample k-tuples of k elements from the set 1:n
}
\examples{
random_combs(5,3,10)
}
