% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{sample_from_tree}
\alias{sample_from_tree}
\title{Samples data from a Gaussian latent tree model}
\usage{
sample_from_tree(tree, m, n)
}
\arguments{
\item{tree}{An igraph object that is a tree. 
It is assumed that the first \code{m} nodes correspond to oberseved nodes and 
that model parameters are given in the following way:
\code{V(g)$var} is the variance of the observed variables and 
that \code{E(g)$corr} are the edge correlations.}

\item{m}{Integer, number of observed nodes.}

\item{n}{Integer, number of samples being sampled.}
}
\value{
Matrix with sampled data with m columns and n rows. 
Each row corresponds to one sample.
}
\description{
Samples data from a Gaussian latent tree model
}
\examples{
# Create tree
vertices <- data.frame(name=seq(1,8), type=c(rep(1,5), rep(2,3))) # 1=observed, 2=latent
edges <- data.frame(from=c(1,2,3,4,5,6,7), to=c(8,8,6,6,7,7,8))
tree <- graph_from_data_frame(edges, directed=FALSE, vertices=vertices)

# Sample data from tree
V(tree)$var = rep(1,8)
E(tree)$corr = rep(0.7,7)
sample_from_tree(tree, m=5, n=500)
}
