\name{TestGLTM-package}
\alias{TestGLTM-package}
\alias{TestGLTM}
\docType{package}
\title{
  Testing Gaussian Latent Tree Models
}
\description{
  Tests the goodness-of-fit of Gaussian latent tree models. Three algebraic testing methodologies are implemented as well as the likelihood ratio test using the EM-algorithm. Supports all Gaussian latent tree models where the observed variables correspond to the leaves of the tree.
}
%\details{
%  This section should provide a more detailed overview of how to use the
%  package, including the most important functions.
%&}
\author{
Nils Sturma

Maintainer: Nils Sturma <nils.sturma@tum.de>
}
%\keyword{ package }
%\examples{
%  \dontrun{
%     ## Optional simple examples of the most important functions
%     ## These can be in \dontrun{} and \donttest{} blocks.   
%  }
%}
